Add fetch-refs.js, shared between WICG and WHATWG.
