+++ refs/w3c.json
@@ -26066,7 +26066,7 @@
         ],
         "href": "http://www.w3.org/TR/html-aria/",
         "title": "ARIA in HTML",
-        "rawDate": "2016-03-10",
+        "rawDate": "2016-05-11",
         "status": "WD",
         "publisher": "W3C",
         "deliveredBy": [
@@ -26193,6 +26193,20 @@
                 "href": "http://www.w3.org/TR/2016/WD-html-aria-20160310/",
                 "title": "ARIA in HTML",
                 "rawDate": "2016-03-10",
+                "status": "WD",
+                "publisher": "W3C",
+                "deliveredBy": [
+                    "http://www.w3.org/WebPlatform/WG/"
+                ],
+                "source": "https://www.w3.org/2002/01/tr-automation/tr.rdf"
+            },
+            "20160511": {
+                "authors": [
+                    "Steve Faulkner"
+                ],
+                "href": "http://www.w3.org/TR/2016/WD-html-aria-20160511/",
+                "title": "ARIA in HTML",
+                "rawDate": "2016-05-11",
                 "status": "WD",
                 "publisher": "W3C",
                 "deliveredBy": [
