+++ test/ref-list.json
@@ -2809,6 +2809,7 @@
     "HTML-ARIA-20151216",
     "HTML-ARIA-20160115",
     "HTML-ARIA-20160310",
+    "HTML-ARIA-20160511",
     "HTML-BIDI",
     "HTML-BIDI-20100304",
     "HTML-BIDI-20150721",
@@ -13529,6 +13530,7 @@
     "RFC7828",
     "RFC7829",
     "RFC783",
+    "RFC7830",
     "RFC7834",
     "RFC7835",
     "RFC7836",
@@ -13536,6 +13538,7 @@
     "RFC784",
     "RFC7840",
     "RFC7842",
+    "RFC7843",
     "RFC7845",
     "RFC7847",
     "RFC785",
@@ -19668,6 +19671,7 @@
     "html-aria-20151216",
     "html-aria-20160115",
     "html-aria-20160310",
+    "html-aria-20160511",
     "html-bidi",
     "html-bidi-20100304",
     "html-bidi-20150721",
@@ -29619,6 +29623,7 @@
     "rfc7828",
     "rfc7829",
     "rfc783",
+    "rfc7830",
     "rfc7834",
     "rfc7835",
     "rfc7836",
@@ -29626,6 +29631,7 @@
     "rfc784",
     "rfc7840",
     "rfc7842",
+    "rfc7843",
     "rfc7845",
     "rfc7847",
     "rfc785",
